fix(core): address review on the stale-refresh guard
- use the overridable `validateNode` for the staleness check, keeping it in
  sync with the entry check in `refresh`
- move the two specs to `tree-consistency.spec.ts`, which is the home for
  root-swap-during-resolve scenarios, so `tree.spec.ts` is untouched
