+++ packages/core/src/browser/tree/tree.spec.ts
@@ -15,30 +15,11 @@
 // *****************************************************************************
 
 import * as assert from 'assert';
-import { injectable } from 'inversify';
-import { TreeNode, CompositeTreeNode, TreeImpl } from './tree';
+import { TreeNode, CompositeTreeNode } from './tree';
 import { TreeModel } from './tree-model';
 import { MockTreeModel } from './test/mock-tree-model';
 import { expect } from 'chai';
 import { createTreeTestContainer } from './test/tree-test-container';
-import { Deferred, timeout } from '../../common/promise-util';
-import { ILogger } from '../../common';
-import { MockLogger } from '../../common/test/mock-logger';
-
-/**
- * A `TreeImpl` whose `resolveChildren` can be stalled per node id, to simulate
- * slow child resolution (e.g. a file system request) in tests.
- */
-@injectable()
-class ControllableTree extends TreeImpl {
-
-    readonly resolveRequests = new Map<string, Deferred<TreeNode[]>>();
-
-    protected override resolveChildren(parent: CompositeTreeNode): Promise<TreeNode[]> {
-        const pending = this.resolveRequests.get(parent.id);
-        return pending ? pending.promise : super.resolveChildren(parent);
-    }
-}
 
 describe('Tree', () => {
 
@@ -244,58 +225,6 @@ describe('Tree', () => {
                 expect(expectedRefreshedNodes.size).to.be.equal(0);
                 done();
             });
-        });
-    });
-
-    describe('refresh - stale refreshes', () => {
-
-        let tree: ControllableTree;
-        let loggedErrors: unknown[][];
-
-        beforeEach(async () => {
-            const container = createTreeTestContainer();
-            container.rebind(TreeImpl).to(ControllableTree);
-            model = container.get(TreeModel);
-            tree = container.get(TreeImpl) as ControllableTree;
-            loggedErrors = [];
-            const logger = container.get<MockLogger>(ILogger);
-            logger.error = async (...args: unknown[]) => { loggedErrors.push(args); };
-            model.root = MockTreeModel.HIERARCHICAL_MOCK_ROOT();
-            // let the refresh cascade triggered by setting the root settle
-            await timeout(0);
-        });
-
-        it('drops an in-flight refresh when the root is replaced meanwhile', async () => {
-            const stale = retrieveNode<CompositeTreeNode>('1.2');
-            const gate = new Deferred<TreeNode[]>();
-            tree.resolveRequests.set('1.2', gate);
-            const pendingRefresh = tree.refresh(stale);
-            tree.resolveRequests.delete('1.2');
-
-            model.root = MockTreeModel.HIERARCHICAL_MOCK_ROOT();
-            gate.resolve(Array.from(stale.children));
-
-            expect(await pendingRefresh).to.be.undefined;
-            expect(loggedErrors).to.have.lengthOf(0);
-            const fresh = model.getNode('1.2');
-            expect(fresh).to.not.be.undefined;
-            expect(fresh).to.not.equal(stale);
-        });
-
-        it('does not resurrect a removed subtree when a stale refresh completes', async () => {
-            const target = retrieveNode<CompositeTreeNode>('1.2');
-            const gate = new Deferred<TreeNode[]>();
-            tree.resolveRequests.set('1.2', gate);
-            const pendingRefresh = tree.refresh(target);
-            tree.resolveRequests.delete('1.2');
-
-            CompositeTreeNode.removeChild(target.parent as CompositeTreeNode, target, tree);
-            expect(model.getNode('1.2')).to.be.undefined;
-            gate.resolve(Array.from(target.children));
-
-            expect(await pendingRefresh).to.be.undefined;
-            expect(model.getNode('1.2')).to.be.undefined;
-            expect(model.getNode('1.2.1')).to.be.undefined;
         });
     });
