+++ packages/core/src/browser/tree/tree.ts
@@ -354,7 +354,7 @@ export class TreeImpl implements Tree {
                 result = parent;
                 const children = await this.resolveChildren(parent);
                 if (cancellationToken?.isCancellationRequested) { return; }
-                if (this.getNode(parent.id) !== parent) {
+                if (this.validateNode(parent) !== parent) {
                     // the tree changed while children were being resolved (e.g. a new root was
                     // set or the node was removed): drop the stale refresh
                     return undefined;
